# api/v1alpha1/resource_manager/types.gen.go
// Package resource_manager provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package resource_manager

import (
	"time"
)

// Error RFC 7807 compliant error response
type Error struct {
	// Detail Human-readable explanation specific to this occurrence
	Detail *string `json:"detail,omitempty"`

	// Instance URI reference for this specific error occurrence
	Instance *string `json:"instance,omitempty"`

	// Status HTTP status code
	Status *int `json:"status,omitempty"`

	// Title Short human-readable summary of the problem
	Title string `json:"title"`

	// Type URI reference identifying the error type
	Type string `json:"type"`
}

// Health Health status singleton resource
type Health struct {
	// Path Canonical path of the resource
	Path *string `json:"path,omitempty"`

	// Status Health status
	Status *string `json:"status,omitempty"`
}

// ServiceTypeInstance Full service type instance resource representation
type ServiceTypeInstance struct {
	// CreateTime Timestamp when the instance was first created
	CreateTime *time.Time `json:"create_time,omitempty"`

	// Id Unique identifier for the Service Type Instance
	Id *string `json:"id,omitempty"`

	// Path Resource path identifier
	Path *string `json:"path,omitempty"`

	// ProviderName Name of the provider
	ProviderName string `json:"provider_name"`

	// Spec Service specification following one of the supported service type
	// schemas (VMSpec, ContainerSpec, DatabaseSpec, or ClusterSpec).
	Spec map[string]interface{} `json:"spec"`

	// Status Status of the instance
	Status *string `json:"status,omitempty"`

	// UpdateTime Timestamp when the instance was last updated
	UpdateTime *time.Time `json:"update_time,omitempty"`
}

// ServiceTypeInstanceList Paginated list of instances
type ServiceTypeInstanceList struct {
	Instances *[]ServiceTypeInstance `json:"instances,omitempty"`

	// NextPageToken Token for retrieving the next page of results
	NextPageToken *string `json:"next_page_token,omitempty"`
}

// InstanceIdPath defines model for InstanceIdPath.
type InstanceIdPath = string

// ListInstancesParams defines parameters for ListInstances.
type ListInstancesParams struct {
	// Provider Filter service provider
	Provider *string `form:"provider,omitempty" json:"provider,omitempty"`

	// MaxPageSize Maximum number of results per page
	MaxPageSize *int `form:"max_page_size,omitempty" json:"max_page_size,omitempty"`

	// PageToken Token for pagination
	PageToken *string `form:"page_token,omitempty" json:"page_token,omitempty"`
}

// CreateInstanceParams defines parameters for CreateInstance.
type CreateInstanceParams struct {
	// Id Optional ID for service type instance
	Id *string `form:"id,omitempty" json:"id,omitempty"`
}

// CreateInstanceJSONRequestBody defines body for CreateInstance for application/json ContentType.
type CreateInstanceJSONRequestBody = ServiceTypeInstance